fix some type comparison warnings

+++ src/fullgrid/Tensor.hpp
@@ -26,7 +26,7 @@ class TensorIndexer {
       localOffsets_[j] = nrElements;
       nrElements = nrElements * extents_[j];
     }
-    assert(this->size() == nrElements);
+    assert(this->size() == static_cast<size_t>(nrElements));
   }
 
   // have only move constructors for now
@@ -58,7 +58,7 @@ class TensorIndexer {
     }
     auto size = std::accumulate(this->extents_.begin(), this->extents_.end(), 1U,
                                 std::multiplies<size_t>());
-    assert(size < std::numeric_limits<IndexType>::max());
+    assert(size < static_cast<size_t>(std::numeric_limits<IndexType>::max()));
     return size;
   }
